Move invoke_training/lora to invoke_training/core/lora.
