+++ src/invoke_training/core/lora/injection/lora_layer_collection.py
@@ -2,7 +2,7 @@ import typing
 
 import torch
 
-from invoke_training.lora.layers import BaseLoRALayer
+from invoke_training.core.lora.layers import BaseLoRALayer
 
 
 class LoRALayerCollection(torch.nn.Module):
