+++ src/invoke_training/core/lora/layers/lora_linear_layer.py
@@ -2,7 +2,7 @@ import math
 
 import torch
 
-from invoke_training.lora.layers import BaseLoRALayer
+from invoke_training.core.lora.layers import BaseLoRALayer
 
 
 class LoRALinearLayer(BaseLoRALayer):
