+++ src/invoke_training/training/dreambooth_lora/dreambooth_lora_sd.py
@@ -11,7 +11,7 @@ from diffusers.optimization import get_scheduler
 from tqdm.auto import tqdm
 from transformers import CLIPTextModel
 
-from invoke_training.lora.injection.stable_diffusion import (
+from invoke_training.core.lora.injection.stable_diffusion import (
     inject_lora_into_clip_text_encoder,
     inject_lora_into_unet,
 )
