+++ src/invoke_training/training/shared/lora_checkpoint_utils.py
@@ -2,7 +2,7 @@ import logging
 
 import torch
 
-from invoke_training.lora.injection.stable_diffusion import (
+from invoke_training.core.lora.injection.stable_diffusion import (
     convert_lora_state_dict_to_kohya_format,
 )
 from invoke_training.training.shared.checkpoint_tracker import CheckpointTracker
